Move common classes to Examples Cdap module
