+++ examples/java/cdap/zendesk/src/main/java/org/apache/beam/examples/complete/cdap/zendesk/utils/PluginConfigOptionsConverter.java
@@ -15,13 +15,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.complete.cdap.utils;
+package org.apache.beam.examples.complete.cdap.zendesk.utils;
 
 import io.cdap.plugin.common.Constants;
 import io.cdap.plugin.zendesk.source.batch.ZendeskBatchSourceConfig;
 import io.cdap.plugin.zendesk.source.common.config.BaseZendeskSourceConfig;
 import java.util.Map;
-import org.apache.beam.examples.complete.cdap.options.CdapZendeskOptions;
+import org.apache.beam.examples.complete.cdap.zendesk.options.CdapZendeskOptions;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 
 /**
